fix and split tests
